fix: when taking struct fields they should be merged into the output in the correct order (#3277)
In various situations we need to fetch some fields from a struct and
then later add more fields for the struct. For example, maybe we have a
`struct<big_string: string, filter: i32>`. We might query with a filter
on `filter` and then use late materialization to take `big_string`. When
we do this we were previously creating `struct<filter: i32, big_string:
string>` which would cause issues since that isn't the correct data
type.

In creating this fix I added a new `Projection` concept and I would like
to slowly replace a lot of the places where we use schemas as
projections to use `Projection` instead. Not necessarily for performance
but more for convenience.