feat: cache btree sub-index pages (#3309)
The btree currently only caches the the lookup. This means there will
always need to be at least one IOP per search 4Ki rows that match the
filter to load the sub-index (flat) data.

This PR adds a cache for the data pages as well.

The cache defaults to 512MiB and is technically configurable via
`LANCE_BTREE_CACHE_SIZE` but this is just a stop-gap. We should ideally
have a single size limit for all scalar indices (or, even better, a
single size limit for ALL indices). However, this is a bit tricky. I
don't think moka caches can share capacity across several caches and
using a single moka cache for all places we need cache is tricky because
moka caches are typed (in some places we use `Arc<dyn Any>` which might
be a possible solution though we still need to deal with the fact that
we key the different caches by `u32`, `String`, and
`object_store::Path`).

We can resolve this technical debt in
https://github.com/lancedb/lance/issues/3136