feat(java): support statistics row num for lance scan (#3304)
Support statistics row num for lance scan, and with this statistics the
spark will choose the broadcast to join for a small table.
But now the byte size of lance dataset is inferred by row number. It is
not very precise.
Maybe We should store the file size in meta as disscus in #3221.