feat: move fsl handling to structural encodings and add support for miniblock (#3324)
Previously we had some support for FSL in the full zip encoder by
treating FSL as a "compression". Implementing this the same way in
miniblock was rather tricky and it seemed more correct to handle FSL at
the structural layer.

Now the structural encoding flattens data and the compression layer just
sees a larger array of items and is unaware of any FSL. We _may_ need to
revisit this decision in the future if we want to add a "vector
compression" algorithm but, as none exists yet, we can worry about that
later.