fix: avoid double-take in some scenarios (#3357)
The code was a little over-eager with regards to eager materialization.
I think this was just a mistake from a recent refactor. This led to a
double-take (also in the code). In most cases this double-take would be
smoothed out by the double-take reducing optimization step. However, in
some cases (e.g. if there was a limit between the takes) the takes
weren't being combined.

We _could_ improve the double-take optimization step to pushdown through
a limit node but it was an easier to fix to avoid accidentally
generating the double-take in the first place.

I added a test case for this in test_plans although that test case
highlights another interesting optimization we could apply which is to
push limits down into vector searches.