feat: make it possible to build lance without protoc (except on Windows) (#3363)
Inspired by https://github.com/substrait-io/substrait-validator/pull/356
this PR adds a `protoc` feature to `lance`. If the feature is specified
then we will use
[`protobuf-src`](https://github.com/MaterializeInc/rust-protobuf-native)
to build a vendored copy of `protoc`.

This increases build times slightly (need to compile `protoc` as part of
the build) but removes the need for an external copy of `protoc`.

At the moment it is not possible to build both the `protoc` and
`substrait` features because `datafusion-substrait` does not yet have
its own `protoc` feature (but it will hopefully have one added soon).