fix: ensure that 'block_size' parameter is properly propagated in the ObjectStore (#3403)
Previously, setting a block_size parameter did not change the range
coalescing being done in FileScheduler's [submit_request
](https://github.com/lancedb/lance/blob/main/rust/lance-io/src/scheduler.rs#L717)
function. This was because the FileScheduler was using the block_size
parameter in the underlying ObjectStore, but that parameter was not
being properly propagated in the configure_store method.

This change keeps the same defaults but always uses the parameter value
if it is set.

Verified via unit tests and profiling to determine increasing block size
now does actually decrease the number of get_range queries.