feat: finish up variable-length encodings in the full-zip path (#3344)
This adds the last structural path for 2.1, full zip encoding of
variable length data. Scheduling this turned out to be a little trickier
than I had planned. There is no easy way to know where to slice the
fully-zipped buffer when doing decoding. Currently we settle this
problem by unzipping in the indirect scheduling task. There are some
alternative possibilities that I have documented but for now I think
this will be good enough and we can iterate on this going forwards.