chore: add binary array generator that generates different sized binary items (#3390)
The current generator for binary data always generates data that has the
exact same length. This PR adds a variation that generates binary
elements of different lengths (the lengths are uniformly sampled from a
given range).

The existing generator is renamed to fixedbin and this one takes the
varbin spot.