feat: auto-migrate old index metadata (#3428)
Follow up to https://github.com/lancedb/lance/pull/3377. That PR made
`index_statistics()` error by default. This ended up being a footgun for
some users who rely heavily on that method. So instead of forcing the
user to do the migration themself, we do it for them. It can be disabled
using an environment variable.