fix: don't eagerly materialize fields that the user hasn't asked for (#3442)
We added logic a while back to eagerly materialize fields if they are
narrow and there is a filter. However, we forgot to ensure that those
fields are actually part of the final projection. The result is that we
end up loading many columns the user doesn't want and then throwing them
away.

This fix changes the set of fields we load to only be those that are
asked for.