fix: temporarily disable spilling when training indices on string columns (#3469)
Until we upgrade to the next DF release (46) we cannot rely on spilling
when working with string data. Users continue to get errors unrelated to
the size of the spill pool or the amount of data they have. This
disables spilling entirely on string columns (which is the typical
workaround) until we get a stable solution.