feat: add support for ngram indices (#3468)
Ngram indices are indices that can speed up various string filters. To
start with they will be able to speed up `contains(col, 'substr')`
filters. They work by creating a bitmap for each ngram (short sequence
of characters) in a value. For example, consider an index of 1-grams.
This would create a bitmap for each letter of the alphabet. Then, at
query time, we can use this to narrow down which strings could
potentially satisfy the query.

This is the first scalar index that requires a "recheck" step. It
doesn't tell us exactly which rows satisfy the query. It only narrows
down the list. Other indices that might behave like this are bloom
filters and zone maps. This means that we need to still apply the filter
on the results of the index search. A good portion of this PR is adding
support for this concept into the scanner.

---------

Co-authored-by: Will Jones <willjones127@gmail.com>