fix: prevent despecialization of object store methods (#3506)
Adding enforcement that wrappers implement all ObjectStore methods. If
they don't, then the specialized implementations in the base
`ObjectStore` could be bypasses, causing key optimizations to be missed
out on.

I didn't see any real cases here, but adding this to be cautious.
`delete_stream()` a common one we might worry about it, but it seems
like tracing store already had this.