chore: suppress humantime advisory (#3529)
The `humantime` package is no longer maintained. We should move to avoid
using it. We were using it indirectly via `env_logger` and
`object_store`. This PR updates `env_logger` to the latest version
(which uses `jiff` instead of `humantime`).

`object_store` is one of our core dependencies and we generally update
it as new versions are released. If (and when) `humantime` is replaced
with `jiff` we will remove the dependency then. This may be several
months. https://github.com/apache/arrow-rs/pull/7261 can be used for
tracking.

Until then we should suppress the advisory.