perf: coalesce continuous indices into ranges if possible (#3513)
In `DecodeBatchScheduler`, when performing `schedule_take` with a given
list of indices, the current implementation generates a list of ranges,
each containing a single index. However, in cases where the indices are
continuous, we can merge them into a single range instead of multiple
separate ranges. This optimization improves efficiency, particularly
benefiting dense queries that return most of the records in the dataset,
as demonstrated in our benchmarks.