fix: work around deranged breaking change not labeled as such (#3591)
A breaking change was introduced in
https://github.com/jhpratt/deranged/issues/18 which was not given a
semver bump. As a result we pick it up and it fails the "no lock file"
test.

This PR just avoids the try_into entirely since it doesn't seem to be
necessary (we only work on 64-bit machines so usize->u64 should be
safe).