chore: remove SNAPSHOT from version (#3600)
Reverts part of #3546 which added `-SNAPSHOT` to the versions. Currently
the CI build system does not publish Java artifacts on pre-releases.
There is also nothing in the build script to remove the `-SNAPSHOT`
designation from the version. As a result the publish failed.

Currently, CI requires the version specifier point to the next stable
version that will be released. This restores that so the next stable
release can succeed.