perf: migrate to `ManifestLocation`, add e_tag (#3592)
* Migrates all methods of `CommitHandler` to just use
`ManifestLocation`.
* Eliminates `O(num_manifests)` IOPS from `cleanup_old_versions`, since
we no longer have to make a separate `HEAD` request to get the size of
the file.
* Eliminates `O(num_manifests)` IOPS from `list_versions()`, similar
reasons as above.
* Adds `e_tag` to `ManifestLocation`, so we can check we are loading the
expected manifest. This eliminates the possibility that we are caching
an old version of the manifest, in cases where the dataset has been
deleted and recreated to the same version number.