feat: int8 support for distance functions (#3605)
Using
[AllMiniLM12V2](https://huggingface.co/sentence-transformers/all-MiniLM-L12-v2)
to generate embeddings, I found that with a tiny bit of transformation,
I can store the embedding as int8 to massively reduce its storage
requirement and turn it into f32 during distance calculation (using
cosine and dot) with marginal change to the outcome.

I would like to add support for int8 by converting FixedSizeList type of
Int8Array to Float32Array during calculation.

The conversion is done with a ``convert_to_floating_point`` on the
``FixedSizeListArrayExt`` trait. In theory, this can also work with
other types like uint8, int32, int64. This would make a big difference
in storage requirement on the compatible models.

---------

Co-authored-by: LuQQiu <luqiujob@gmail.com>