feat: support AND query and improve phrase query performance (#3623)
this also improves the phrase query performance by skipping documents
that's impossible to be a candidate.
before we always set the WAND threshold to 0, and update the threshold
when hit the limit. But for AND query, only documents that contain all
terms can be a candidate, so the init threshold can be the sum of upper
bounds

this also refines the pyo3 APIs so that we can get rid of that
complicated `parse_fts_query` method

---------

Signed-off-by: BubbleCal <bubble-cal@outlook.com>