feat: add a function to prewarm an index (#3697)
Querying a cold index can incur a cold-start penalty. In some cases,
users know their indices fit into RAM, and they want to avoid this cold
start penalty by pre-warming the index before it is ever queried. This
PR adds a function `prewarm_index` to make that possible.

In order to make it possible to test this function this PR also adds a
`scan_stats_callback` which can be provided to the scanner to return the
summary statistics back to python. This way we can check those
statistics in unit tests. This feature may also be useful in general to
users that want to know more about scan performance. It's the same data
reported by analyze_plan but in a form that is more easily accessible
programmatically.

This PR only implements pre-warming for the FTS (inverted) index. Adding
prewarming to other indices will be covered by
https://github.com/lancedb/lance/issues/3698

---------

Signed-off-by: BubbleCal <bubble-cal@outlook.com>
Co-authored-by: BubbleCal <bubble-cal@outlook.com>