fix: don't emit the IO core reservation warning on 2-core systems (#3710)
The default value for LANCE_IO_CORE_RESERVATION is 2. So if users are on
a 2-core system we are always going to warn them. This is generally just
confusing for users and not really something they can do much about.
This changes the check to only emit the warning if the user is modifying
LANCE_IO_CORE_RESERVATION (e.g. it is something other than 2).