ci(java): improve caching for lance-jni (#3702)
Reduce the runtime of Java CI from over 10 minutes to 3 minutes. The
caching configuration in CI requires (1) the working directory to be at
the root, (2) the existence of cargo runtime, and (3) a correct mapping
of the JNI build to the maven target directory, (4) do not clean the
build artifact because that will clean all compiled rust artifacts.