feat: add ExpLinkedList (#3721)
This data structure is useful when we need to append a lot of values and
finally iterate over all of them.
Compared to `Vec<T>`, this data structure doesn't waste much memory like
`Vec<T>`, and doesn't need to copy the data when recap.

This is introduced for FTS indexing process, will apply this in next PR

---------

Signed-off-by: BubbleCal <bubble-cal@outlook.com>