feat: strict overwrite operation (#3712)
Modifies the handling of the Overwrite operation to respect
`read_version` when `num_retries` is set to 0 (strict Overwrite).
A strict Overwrite fails if new commits have been written after
`read_version`.
This allows users to more easily build custom transaction and retry
logic if needed.

The first commit refactors only for code deduplication with no
behavioral changes.

The remainder makes two behavioral changes:
1. Each commit will be tried at least once (previously, commits would
always fail when `num_retries` is set to 0)
2. If the operation is Overwrite and `num_retries` is 0, the commit will
be attempted based on `read_version` instead of attempting to check out
the latest version of the dataset.