fix: flat scan frags not covered by scalar index when doing prefilter search (#3768)
When performing a search (vector or FTS) we can use a scalar index to
help us calculate the prefilter. In some cases we don't even have to
materialize the scalar index. However, our old logic was too loose here
and didn't materialize even when the scalar index was behind the
vector/FTS index (didn't cover all the fragments).

This PR restricts the non-materialization path to only be followed when
the scalar index is fully caught up to the vector/fts index.