feat: split large reads into smaller parallel reads (#3801)
We can often end up with large pages. This is easier in 2.1 than it is
in 2.0 but it can happen in 2.0 as well. When this happens we make one
large request into object storage. However, these large requests are
problematic. Partly because we get less performance (less parallelism)
and partly because they are prone to failure, especially timeout, when
the object store is under stress. As a result, it is better to split
requests into smaller requests.