feat: introduce new FilteredReadExec which will replace the initial scan/take (#3782)
Introduces a new exec node `FilteredReadExec` which will fully replace
`ScanExec` and also replace many instances of `TakeExec` (we will still
use `TakeExec` after a vector / fts search but we could use filtered
read eventually there too if we want to simplify).

The new node has a few advantages over the old behavior and also lays
the ground work for further enhancements:

* Index operations now occur on ranges, and not materialized rows. This
means that a filter like `unique_id != 7757` will not requires an
expensive materialization step.
* Deletions are now materialized before the read. This won't usually
reduce I/O (since we probably just coalesce the small gap in most cases)
but it can in some cases (large gaps / large types) and it also avoids
any decode of the deleted rows so it is slightly cheaper.
* Introduces support for a "scan range before filter" which can be
useful in a few niche scenarios. This can also be used when there is no
filter to pushdown the limit (we already did this). We also lay the
ground work for a "scan range after filter" which can be used to
pushdown the limit, even when there is a filter, provided the filter is
indexed (or at least indexed enough to find enough filtered rows). This
latter feature isn't quite ready yet in this PR though.
* This also moves the re-check inside the filtered read node (instead of
a dedicated `FilterExec` node) which lays the groundwork for eventually
reducing compute by skipping the filter when we have an `AtLeast` /
`AtMost` result from a scalar index (e.g. this can dramatically reduce
compute of filters like `NOT string_contains(my_str, 'some_test_str')`
when there is an ngram index.
* Adds support for partitioned scans which can be used to introduce
parallel compute (although this will destroy the scan order). This is
mostly useful for expensive filters (e.g. string operations) or SQL
queries with expensive downstream (e.g. TPC-H). Basically anytime we are
compute bound and not I/O bound.

This PR does not yet hook the new exec node into the planner so it is
not yet used in practice. However it is tested in place.

When this is plugged in there will be one minor breaking change.
Previously, if there was a large chunk of deleted rows, and this caused
a batch of rows to be loaded that was entirely deleted, we would skip
this batch. I'm not sure we want that behavior. That is because we often
include deleted rows so we can accomplish things like fragment alignment
(e.g. when adding a new column, materialize deleted rows so that we
calculate the right size number of values). With this PR we change that
behavior so that these all-null batches are still included. We still do
not materialize rows for deleted fragments.