fix: logic bug in RowIdTreeMap union (#3811)
During the union operation, if the left hand side contained a partial
bitmap, and the right hand side contained a full bitmap, we were
ignoring the right hand side. The correct thing to do is to replace the
left hand side with a full bitmap.

This was detected via a property test failure on the bitmap union test.
The intersect test also failed but this was a failure in the unit test
and not the implementation.