fix: fix python tracing (#3807)
When we added logging of trace events we made a passthru subscriber that
we could switch at runtime from nothing to chrome. This passthru did not
realize that the `Subscriber` trait has many methods with default impls
and these were not getting passed through. The chrome subscriber doesn't
care about these but the `Registry` subscriber does and these are very
important (the `Registry` is the one that handles the "current span"
concept).

I reworked it so we are always using a `Registry` and now the passthru
point is a `Layer` and not a `Subscriber`. This has a simpler interface
and now tracing context is working as expected.

The one downside is that we need to setup filtering when logging is
initialized which happens as soon as the module is imported. As a
result, we cannot pass the log level in as a separate call. For now we
just utilize the `LANCE_LOG` variable which is probably fine. If we
needed the fine-grain control in the future we could either defer
initialization of logging or make new environment variables for tracing
levels.