fix: allow nulls when blob data files are missing (#3843)
Currently we translate nulls into a struct `{"position": 1, "size": 0}`.
As a result, we had the descriptions schema marked as non-nullable.
However, this is slightly confusing (if user scans blob column they
should get null, not this placeholder struct) and caused issues when
data files were missing (since we populate those with null).

This fixes the latter problem by changing the descriptions schema to
nullable. However, we should also fix the other problem at some point.
We would need to wrap the packed struct decoder we use for blob
descriptions with a layer that converts this placeholder value to null.

That being said, we could probably enact this cleanup in 2.1 and not
worry about it for 2.0 as it is fairly minor (users aren't generally
looking at the blob descriptions).

Closes #3840