fix: update lindera to 0.40.2 and correct outdated URL (#3852)
I'm currently experimenting with full-text search in lancedb using
lindera implemented in https://github.com/lancedb/lance/pull/3218.
So far, it's working well with my local dataset.

During the process, I noticed that the URL for the Lindera dictionary in
the Lance repository was outdated.
This pull request [updates that
URL](https://github.com/lindera/lindera/blob/v0.40.2/lindera-ipadic/build.rs#L11)
and lindera version.

Most of the changes are in the Cargo.toml file.
Since I'm not very familiar with Rust, please feel free to suggest a
better alternative if there's a more appropriate way to handle this.