fix: use `e_tag` in cache key for cached manifests (#3881)
This fixes the failing test like
https://github.com/lancedb/lance/actions/runs/15223146317/job/42821795030

Background: if users drops a dataset, and then re-create it, it's
possible for long-lived readers to have loaded a manifest from the
dataset prior to deletion. We might have identical manifest file names
`1.manifest` and `1.manifest`, but different contents. To detect these
cases, we keep track of the `e_tag` of these. This PR makes sure we use
this `e_tag` in the cache key.