ci: disable torch tests on macos CI (#3853)
The macos CI jobs have been hanging consistently for quite some time.
After pytest finishes the python interpreter hangs on exit. I was able
to extract a stack trace using the `faulthandler` module combined with
`gtimeout`:

```
 Thread 0x00000001edfacf80 (most recent call first):
  File "/Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/multiprocessing/resource_tracker.py", line 111 in _stop_locked
  File Current thread 0x00000001edfacf80 (most recent call first):
  File "/Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/multiprocessing/resource_tracker.py"", line /Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/multiprocessing/resource_tracker.py111" in _stop_locked
  File "/Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/multiprocessing/resource_tracker.py", line 86 in _stop
  File "/Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/multiprocessing/resource_tracker.py", line 77 in __del__
, line 86 in _stop
  File "/Library/Frameworks/Python.framework/Versions/3.12/lib/python3.12/multiprocessing/resource_tracker.py", line 77 in __del__
```

My guess is that somehow the pytorch tests are managing to trigger a
fork and it's related to https://github.com/python/cpython/issues/96971

I tried setting the multiprocessing start method to spawn in
`conftest.py` but that didn't seem to help. For now, I think we need to
get CI passing again and move on. I would like to just disable pytorch
tests on macos.