fix: last item could be dropped from directory listing (#3892)
https://github.com/lancedb/lance/pull/3644 introduced a static listing
stream, with a bug that listing result could be dropped if the sender
has finished sending all items but the receiver is still in pending
state. When this bug happens, the dataset would checkout_latest to an
old version, and when the the dataset is used in commit, writer
continuously tries to commit to an existing version.