fix: prevent and handle empty fragments (#3934)
#3933 

We found it's possible to create datasets with empty fragments, and that
this can break remapping.

This PR:

1. Makes re-mapping gracefully return early when it detects that the
re-map is just all deletions. This avoids the error reported in the
issue.
2. Has the commit step remove any empty fragments. This will fix any
broken datasets on any write transaction.
3. Makes sure that the `TransactionRebase` will delete full fragments if
the merged deletion file covers the whole fragment. This was the source
of the empty fragments.