feat: wrap FilterExec in LanceFilterExec (#3939)
This wraps usage of Datafusion FilterExec in a LanceFilterExec node that
carries the original logical Expr for the filter alongside the wrapped
FilterExec.

The reason for this is to enable a datafusion optimizer rule to more
easily serialize the execution logic of the filter (using substrait) and
send the node over the wire.