feat: rebase conflict between rewrite and frag reuse index cleanup (#3952)
This PR ensures the following combination of operations can correctly
resolve conflicts:
- rewrite with FRI against any normal index update: succeed
- any normal index update against rewrite with FRI: succeed
- rewrite with FRI against rewrite with FRI: fail (could be optimized in
the future to rebase)
- rewrite with FRI against FRI cleanup: succeed after rebase
- FRI cleanup against rewrite with FRI: succeed after rebase
-  FRI cleanup against FRI cleanup: fail

In addition, I did a full refactoring of the content in
`conflict_resolver` and `transaction` to make sure we do not need to
compute some states repeatedly (e.g. modified_fragment_ids) and the
conflict resolution logic all lives inside `conflict_resolver` instead
of being spread across multiple modules. It also removes the redundant
layer of `Compatible`, `Retryable` and `Incompatible` enums purely to be
matched in `check_txn` and then being translated to `Ok` or
`RetryableConflictError` and `ConflictError`.

With this refactoring, developers can easily find where to modify
conflict resolution behavior of any combination of transactions:
- check the rebase logic of A against B: go to function `check_A_txn`
and go to match case for `B`.
- check rebase finish logic of A: go to function `finish_A`

Once we implement cases like rebasing rewrite against update/delete, I
expect some operations in `check_delete_update_txn` or
`finish_delete_update` can be extracted to be shared, but that can be
left to the next PRs.

Closes #3942