chore: bump dirs to 6.0.0 (#3977)
This is a minor PR that bumps `dirs` dependency to v6. Right now we have
`lance-io` that uses `shellexpand` that depends on `dirs 6.0.0` and we
have `lance-index` that uses `dirs 5.0.0` directly. This reduces the
duplicate dependencies.

We still have `dirs 5.0.0` in examples due to hugging face, but I think
that's less important for production code.