fix: don't crash when reading index trained on old version (#3978)
Indices trained with old versions of lance did not record the index type
in the protobuf. In those cases we need to fallback to inferring the
index type from clues. One of our regression benchmarks uses an example
of such a dataset.

It turns out we recently introduced a regression where we were passing
in the wrong value (index name instead of column name) to this
inference. This PR fixes that and introduces a new migration test to
regress the behavior.