fix: handle case in 2.1 writer where lists are masked by null struct (#3946)
In Arrow a null struct has an actual item underneath. If that actual
item was an empty / null list then the repdef utilities were encoding it
as an empty / null list and not a null struct. Unfortunately, the fix
was not simple.

This PR reworks the builders for repdef levels so that we now build them
from outside-to-inside instead of the reverse. This is a refactor that
has probably been overdue for a while. The resulting code is (IMHO)
considerably more intuitive.