refactor: change the default process-wide IOPS limit to be 128 (#3972)
The 2.1 reader has various limits that will cause it to pause. It will
not read too much data at once (the readahead queue) and it will be
limited by the `LANCE_IO_THREADS` variable which controls how many
concurrent requests we allow for a single scan.

There is a further limit, which is a process-wide limit, on the number
of concurrent IOPS. Previously this limit was disabled. However, it is
rather easy to get in a situation where excessive IOPS are generated
when doing any kind of multi-threaded workload. For example, compaction
will trigger this as it spawns multiple compaction tasks and each task
needs to read from object storage.

This PR changes the default to enabled and sets it at 128.