feat!: auto-remap indexes before scan (#3971)
BREAKING CHANGE: this PR uses the fragment reuse index (FRI) to
auto-remap all indexes during their load time. By doing so, an index
will automatically cover the fragments that produced after compaction
even before a physical index remap has happened, and related
scan/FTS/KNN search will be able to leverage it. Note that when we say
"load time", it does not always mean when the index is opened. Most
indexes are loaded on-demand at search time and then cached, this is
when the auto-remap happens for a specific subset of the row IDs in an
index. This change introduces the following breaking changes:
1. add `fri: Option<Arc<FragReuseIndex>>` for all related index loading
functions
2. simplify the FRI index structure. With this auto-remap functionality,
we no longer need to track the old fragment locations to read old
fragments. We now only track fragment ID, physical_rows and
num_deleted_rows of old and new fragments
3. modify signatures of related fragment reuse index functions to adopt
to the new index structure

Note that this change does not include auto-remap for HNSW because it is
not widely used and it is not remapped today anyway. This work is
tracked separately in #3993.

Closes #3837