feat!: move file metadata cache to bytes capacity (#3949)
* Changes `FileMetadataCache` to always use bytes mode. For backwards
compatibility, takes old capacity argument and uses 10MB assumed entry
size to compute actual size.
* Rewrites file metadata cache to be more generic. We will later reuse
`LanceCache` to implement a consolidated index cache. In particular,
uses strings instead of Path for keys.
* Refactored how keys are generated: Dataset now keeps a
`metadata_cache` that inserts with a prefix of the dataset URI. This way
all use of this cache is namespaced, making the cache more resilient to
collisions between datasets.