refactor: move 2.0 code into its own module (#4046)
Creates a v2 module and moves all 2.0 specific code to there. This will
make it easier to understand the code. There were no meaningful changes
during this refactor, code was merely moved around, some deprecated
stuff was removed, and there were some minor renames (no renames to
protobuf).

In addition, this gets rid of lance-encoding-datafusion as we are no
longer going with the zone map approach (and instead treating zone maps
as a standalone scalar index).

Closes #3342