feat: introduce MemWAL index and system index (#4069)
This PR introduces MemTable and WAL to Lance. It:
- includes the MemWAL index metadata structure and all the necessary
operations that are performed against it.
- introduces a new `mem_wal_to_flush` parameter in the `MergeInsertJob`
API so that a process can atomically mark a MemWAL as flushed as well as
updating data.
- introduces a `UpdateMemWalState` transaction type which is used to
resolve concurrent actions to transtion MemWAL states.

In general, the Lance table itself will not directly contains a MemTable
or WAL even after everything is merged, but this index records the
necessary metadata so that each commit to WAL and any state change of a
MemTable from a writer is synchronized through this component.

One main reason I am introducing this as another index is that now we
can officially introduce a category of indexes called "system index". A
system index records some temporary state of the Lance table to improve
the performance of certain use cases. So Fragment Reuse Index is a
system index that helps the read after compaction case, and this MemWAL
index is for fast upsert + read after upsert.

Closes #3986 #3987