perf: use size prefixing for zstd buffer compressor for better decompressing performance (#4029)
# Summary

This PR addresses
[lancedb/lance#4028](https://github.com/lancedb/lance/issues/4028) by
introducing a length-prefixed format for the Zstd buffer compressor. The
goal is to improve the decompression performance of Zstd-compressed
buffers.

# Solution

Currently, the Zstd buffer compressor uses the standard compression API
and decompresses data using the Zstd streaming API. While functional,
this approach is not optimal as the streaming decompression introduces
overhead.

In this PR, I switch to a block decompression approach by:
* Prefixing each compressed buffer with its original (uncompressed)
length (encoded as a 64-bit little-endian integer)
* Using Zstd’s block decompression API during decoding, which avoids the
overhead of stream mode.
* Leveraging the Zstd magic number and a reserved bit in the Zstd frame
header descriptor (as defined in RFC 8878) to distinguish between a
standard Zstd stream and the new length-prefixed format, so that
compatibility can be guaranteed

# Consistency with Existing Approaches
* This technique is consistent with the one used by the [Apache Arrow
IPC format][1][2], which also embeds the uncompressed size to enable
faster, more efficient decompression.
* The LZ4 buffer compressor in lance already follows a similar approach,
as its API natively supports it via the `prepend_size` parameter in the
`compress_to_buffer` function.

# Benchmark Results

In limited benchmarks across several input sizes:

* Decompression speed improved by 30% to 200%, depending on the data.
* Compression speed remains unchanged.
* Compressed size increases by 8 bytes (due to the uncompressed size
prefix).

# References
* [1] Arrow IPC writer:
https://github.com/apache/arrow/blob/bd31f83aaa93db8427bc90285658e29d79ce5efd/cpp/src/arrow/ipc/writer.cc#L230-L233
* [2] Arrow IPC reader:
https://github.com/apache/arrow/blob/bd31f83aaa93db8427bc90285658e29d79ce5efd/cpp/src/arrow/ipc/reader.cc#L398-L407